requirements.txt
# Core simulation
OpenDSSDirect.py>=0.9.0

# Numerical computation
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Optimization
cvxpy>=1.4.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Config & I/O
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
